+++ src/main/java/com/pixelsapphire/wanmin/data/Storage.java
@@ -10,11 +10,11 @@ import java.util.Date;
 
 public class Storage {
     private final @NotNull Product product;
-    private final int amount;
+    private final float amount;
     private Date terminDue;
     private final @NotNull ForeignInvoice invoice;
 
-    private Storage(@NotNull Product product, int amount, Date terminDue, @NotNull ForeignInvoice invoice) {
+    private Storage(@NotNull Product product, float amount, Date terminDue, @NotNull ForeignInvoice invoice) {
         this.product = product;
         this.amount = amount;
         this.terminDue = terminDue;
@@ -24,7 +24,7 @@ public class Storage {
     @Contract("_, _, _ -> new")
     public static @NotNull Storage fromRecord(@NotNull ResultSet record, @NotNull Provider<Product> productProvider,
                                               @NotNull Provider<ForeignInvoice> invoiceProvider) throws SQLException {
-        return new Storage(productProvider.getByValue(record.getString("produkt")), record.getInt("ilosc"),
+        return new Storage(productProvider.getByValue(record.getString("produkt")), record.getFloat("ilosc"),
                 record.getDate("data_waznosci"), invoiceProvider.getByValue(record.getString("faktura")));
     }
 
@@ -32,7 +32,7 @@ public class Storage {
         return terminDue;
     }
 
-    public int getAmount() {
+    public float getAmount() {
         return amount;
     }
